+++ Autopark/Staff.cpp
@@ -2,6 +2,7 @@
 #include <string>
 #include <sqlite3.h>
 #include <iomanip>
+#include <Format.h>
 
 using namespace std;
 
@@ -28,8 +29,8 @@ void staff_split() {
 
 
 void staff_show_row(string id, string a, string b) {
-    cout << VERTICAL_SPLIT_CODE << setfill(' ') << setw(CELL_WIDTH - 3) << id << VERTICAL_SPLIT_CODE
-        << setw(CELL_WIDTH - 2) << a << VERTICAL_SPLIT_CODE << setw(CELL_WIDTH - 2) << b << VERTICAL_SPLIT_CODE << endl;
+    cout << VERTICAL_SPLIT_CODE << setfill(' ') << setw(CELL_WIDTH - 3) << cell_format(id, CELL_WIDTH) << VERTICAL_SPLIT_CODE
+        << setw(CELL_WIDTH - 2) << cell_format(a, CELL_WIDTH) << VERTICAL_SPLIT_CODE << setw(CELL_WIDTH - 2) << cell_format(b, CELL_WIDTH) << VERTICAL_SPLIT_CODE << endl;
 }
 
 void staff_show_header() {
@@ -123,8 +124,12 @@ void removeStaff(sqlite3* db) {
     char* zErrMsg = 0;
     string eid;
     setlocale(LC_ALL, "Russian");
-    cout << "\nВведите id для удаления: ";
+    cout << "\nВведите id для удаления (-1 для отмены): ";
     cin >> eid;
+    if (eid == "-1") {
+        cout << "Удаление отменено.\n";
+        return;
+    }
     string sql = "DELETE from " + STAFF_TABLE_NAME + " WHERE id = " + eid;
     int rc = sqlite3_exec(db, sql.c_str(), NULL, NULL, &zErrMsg);
